feat: add all user controller methods

+++ src/usecases/delete-user/delete-user-use-case.ts
@@ -3,7 +3,11 @@ import { UserRepository } from "../../domain/users/user-repository";
 export class DeleteUserUseCase {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async handle(id: string): Promise<void> {
+  async handle(id: string): Promise<void | null> {
+    const user = await this.userRepository.findByID(id);
+
+    if(!user) return null;
+    
     await this.userRepository.delete(id);
   }
 }
